pyproject.toml
[tool.poetry]
name = "delivery_service"
version = "0.1.0"
description = "Delivery Service API for registering parcels, listing types, and viewing parcel details."
authors = ["Alexandr3008 <soloshenko3008@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.15"
psycopg2-binary = "^2.9"
redis = "^5.0"
requests = "^2.31"
django-redis = "^5.4"
celery = "^5.3"
python-dotenv = "^1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-django = "^4.5"
mock = "^5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"